spanconfig: move testing helper out of base package
Pure code movement change.

Release note: None
